refactor: Make Hooks directory and move useSearch
limit is defined in Search page component
